refactor: extract find-address copy module and add missing testIDs (OK-55653)
